adding Perl::Critic::Periscope::ProhibitDollarStar
This policy prohibits the use of the deprecated "$*" variable. I
started with it not because that's a common problem, but because it's
a really easy policy to write and it will help me prove the concept of
this distribution of policies.

I may have to move this file up one directory, out of the "Policy"
directory. I'm not really sure where they're best placed.

The test suite does not work yet. Perl::Critic has a pretty advanced
test suite that builds pieces of documents out of .run files. I'll
need to steal their work somehow. I'll also probably steal some other
of their tests.